backend/requirements.txt
appnope==0.1.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.4
django-allauth==0.38.0
django-cors-headers==3.3.0
django-dotenv==1.4.2
# django-elasticsearch-dsl==0.5.1
djangorestframework==3.9.2
# elasticsearch==6.3.1
# elasticsearch-dsl==6.1.0
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
json-logging-py==0.2
numpy==1.17.1
oauthlib==3.0.0
pandas==0.25.1
pexpect==4.7.0
psycopg2-binary==2.8.4
ptyprocess==0.6.0
python-dateutil==2.7.5
python-json-logger==0.1.10
python3-logstash==0.4.80
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
zappa==0.51.0
